backend/Python/requirements.txt
Django
django-cors-headers
djangorestframework
deepface
